requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-analytics-data>=0.18.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0 